intermediate

+++ internal/domain/service.go
@@ -1,7 +1,7 @@
 package domain
 
 type Time struct {
-	Time int `json:"time" example:"1"`
+	Time int64 `json:"time" example:"1"`
 }
 
 type Site struct {
